GS workspaces and rules are request-specific
