+++ src/layman/geoserver.py
@@ -1,10 +1,12 @@
 import json
 import os
+import re
+
+from flask import g, current_app
 
 from .filesystem import get_user_dir
 from .http import LaymanError
 from .settings import *
-from .util import get_layman_rules, get_non_layman_workspaces
 
 headers_json = {
     'Accept': 'application/json',
@@ -17,46 +19,51 @@ headers_xml = {
 
 
 def get_all_workspaces():
-    r = requests.get(
-        LAYMAN_GS_REST_WORKSPACES,
-        # data=json.dumps(payload),
-        headers=headers_json,
-        auth=LAYMAN_GS_AUTH
-    )
-    r.raise_for_status()
-    # app.logger.info(r.text)
-    all_workspaces = r.json()['workspaces']['workspace']
-    return all_workspaces
+    key = 'layman.geoserver.workspaces'
+    if key not in g:
+        current_app.logger.info('workspaces NOT EXIST')
+        r = requests.get(
+            LAYMAN_GS_REST_WORKSPACES,
+            # data=json.dumps(payload),
+            headers=headers_json,
+            auth=LAYMAN_GS_AUTH
+        )
+        r.raise_for_status()
+        # app.logger.info(r.text)
+        all_workspaces = r.json()['workspaces']['workspace']
+        g.setdefault(key, all_workspaces)
+    else:
+        current_app.logger.info('workspaces EXISTS')
+
+    return g.get(key)
+
 
 def get_all_rules():
-    r = requests.get(
-        LAYMAN_GS_REST_SECURITY_ACL_LAYERS,
-        # data=json.dumps(payload),
-        headers=headers_json,
-        auth=LAYMAN_GS_AUTH
-    )
-    r.raise_for_status()
-    # app.logger.info(r.text)
-    all_rules = r.json()
-    return all_rules
+    key = 'layman.geoserver.rules'
+    if key not in g:
+        r = requests.get(
+            LAYMAN_GS_REST_SECURITY_ACL_LAYERS,
+            # data=json.dumps(payload),
+            headers=headers_json,
+            auth=LAYMAN_GS_AUTH
+        )
+        r.raise_for_status()
+        # app.logger.info(r.text)
+        all_rules = r.json()
+        g.setdefault(key, all_rules)
 
-def classify_workspaces(all_workspaces, all_rules):
-    layman_rules = get_layman_rules(all_rules)
-    non_layman_workspaces = get_non_layman_workspaces(all_workspaces,
-                                                      layman_rules)
-    layman_workspaces = filter(lambda ws: ws not in non_layman_workspaces,
-                               all_workspaces)
-    return layman_workspaces, non_layman_workspaces
+    return g.get(key)
 
-def check_username(username, all_workspaces, all_rules):
-    _, non_layman_workspaces = classify_workspaces(all_workspaces, all_rules)
+
+def check_username(username):
+    non_layman_workspaces = get_non_layman_workspaces()
     if any(ws['name'] == username for ws in non_layman_workspaces):
         raise LaymanError(12, {'workspace': username})
 
+
 def ensure_user_workspace(username):
     all_workspaces = get_all_workspaces()
-    all_rules = get_all_rules()
-    check_username(username, all_workspaces, all_rules)
+    check_username(username)
     if not any(ws['name'] == username for ws in all_workspaces):
         r = requests.post(
             LAYMAN_GS_REST_WORKSPACES,
@@ -247,9 +254,7 @@ def generate_layer_thumbnail(username, layername):
 
 
 def get_layer_info(username, layername):
-    all_workspaces = get_all_workspaces()
-    all_rules = get_all_rules()
-    check_username(username, all_workspaces, all_rules)
+    check_username(username)
 
     try:
         r = requests.get(
@@ -279,9 +284,7 @@ def get_layer_info(username, layername):
 
 
 def get_layer_names(username):
-    all_workspaces = get_all_workspaces()
-    all_rules = get_all_rules()
-    check_username(username, all_workspaces, all_rules)
+    check_username(username)
 
     try:
         r = requests.get(
@@ -300,3 +303,34 @@ def get_layer_names(username):
         return layernames
     except:
         return []
+
+
+def get_layman_rules(all_rules=None, layman_role=LAYMAN_GS_ROLE):
+    if all_rules==None:
+        all_rules = get_all_rules()
+    re_role = r".*\b" + re.escape(layman_role) + r"\b.*"
+    result = {k: v for k, v in all_rules.items() if re.match(re_role, v)}
+    return result
+
+
+def get_non_layman_workspaces(all_workspaces=None, layman_rules=None):
+    if all_workspaces==None:
+        all_workspaces = get_all_workspaces()
+    if layman_rules==None:
+        layman_rules = get_layman_rules()
+    result = [
+        ws for ws in all_workspaces
+        if next((
+            k for k in layman_rules
+            if re.match(r"^" + re.escape(ws['name']) + r"\..*", k)
+        ), None) is None
+    ]
+    return result
+
+
+def get_layman_workspaces():
+    all_workspaces = get_all_workspaces()
+    non_layman_workspaces = get_non_layman_workspaces()
+    layman_workspaces = filter(lambda ws: ws not in non_layman_workspaces,
+                               all_workspaces)
+    return layman_workspaces
