+++ src/layman/util.py
@@ -3,8 +3,6 @@ import unicodedata
 
 from unidecode import unidecode
 
-from .settings import *
-
 
 def slugify(value):
     value = unidecode(value)
@@ -23,17 +21,3 @@ def to_safe_layer_name(value):
     return value
 
 
-def get_layman_rules(all_rules, layman_role=LAYMAN_GS_ROLE):
-    re_role = r".*\b" + re.escape(layman_role) + r"\b.*"
-    result = {k: v for k, v in all_rules.items() if re.match(re_role, v)}
-    return result
-
-def get_non_layman_workspaces(all_workspaces, layman_rules):
-    result = [
-        ws for ws in all_workspaces
-        if next((
-            k for k in layman_rules
-            if re.match(r"^" + re.escape(ws['name']) + r"\..*", k)
-        ), None) is None
-    ]
-    return result
